[Tools] Add github-merge.py script
This script is used when merging pull requests to verify the diff, as
well as the title length and the presence of (ut)ACK comments/reviews.

The resulting merge commit message is better formatted and conducive to
allowing for scripted change logs based on pull requests rather than
individual commits.