Limit iPad support
due to an issue with hiding they pin entry keyboard and not being able
to get the keyboard back, the quickest fix is to limit support for iPad
devices.

The wallet can still technically run on an iPad with this change, but it
will use the iPhone compatibility mode (ugly, but works).

I've also bumped the build version here so it can be pushed to Apple's
App Store Connect quicker.